+++ config/database.php
@@ -1,0 +1,8 @@
+<?php
+return [
+    'sqlite_testing' => [
+        'driver'   => 'sqlite',
+        'database' => ':memory:',
+        'prefix'   => '',
+    ],
+];