+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ## 1.0.0 - 2021-10-05
 
 - initial release
+
+## 2.0.0 - 2021-10-10
+
+- **ANONYMIZABLE** constant definition changed from string to array
+- **db:anonymize** command now can be run with specific model and columns definition
+- Added tests